# tests/Formbase.Host.Tests/Contracts/served-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Formbase.Host | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "{origin}"
    }
  ],
  "paths": {
...
    "/formtypes/{type}/records": {
      "get": {
        "tags": [
          "RecordEndpoints"
        ],
        "summary": "Reads a form type\u0027s projected records",
        "description": "Filters are equality only, one per \u0060filter=column:value\u0060 parameter; the first colon separates the two, so a value may contain colons. Values are compared as the declared column\u0027s type, so \u0060filter=total:42\u0060 matches a number. Ordering takes a comma-separated column list where a leading \u0060-\u0060 reverses that column. Paging is deterministic whether or not the caller orders: the projection\u0027s watermark is appended as a final tie-breaker.",
        "operationId": "QueryRecords",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Formbase-Namespace",
            "in": "header",
            "description": "Which namespace the request is addressed to. Omit it to address the host\u0027s own; naming one this host does not serve is refused rather than answered from the one it does.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordQueryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RecordEndpoints"
        ],
        "summary": "Retires a record",
        "description": "Appends a retirement of the record named by \u0060recordKey\u0060: the projection stops showing it, and its documents stay in the raw stream, where the retirement appears as a document with \u0060retired: true\u0060 and no body. Appending under the key again brings the record back. Nothing checks that the key was ever used. An Idempotency-Key header works as it does for intake.",
        "operationId": "RetireRecord",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Formbase-Namespace",
            "in": "header",
            "description": "Which namespace the request is addressed to. Omit it to address the host\u0027s own; naming one this host does not serve is refused rather than answered from the one it does.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedDocumentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptedDocumentResponse": {
        "required": [
          "documentId",
          "formType"
        ],
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "formType": {
            "type": "string"
          },
          "recordKey": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RecordQueryResponse": {
        "required": [
          "rows",
          "stale"
        ],
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "stale": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}